refs.h change size_t to uint64_t wherever applicable

+++ src/lddmc.c
@@ -316,7 +316,7 @@ lddmc_gc_mark_external_refs(int my_id, int workers)
     if (end >= refs_size) end = refs_size;
 
     // iterate through refs hash table, mark all found
-    size_t *it = refs_iter(first, end);
+    uint64_t *it = refs_iter(first, end);
     while (it != NULL) lddmc_gc_mark_rec(refs_next(&it, end));
 }
